requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
requests>=2.31.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pillow>=10.0.0
bcrypt>=4.0.0
itsdangerous>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
chromadb>=0.4.0
pyjwt>=2.8.0
python-multipart>=0.0.6
